clean-up tests
